Fix for invalid cache key used in the DefaultVotingService
The Default voting service used a cache key that didn't include the
name of the function used to calculate the result value stored in
the cache using the provided cache key.

By doing so, the voting service returned invalid values when
first asking for the count function of a dimension and directly
after that asking for the average of the same dimension.

By including the function name in the cache key, the correct
results are now always returned.